+++ src/BoardHashMap.h
@@ -6,15 +6,15 @@
 
 class BoardHashMap {
 public:
-    bool memorize_board_state(const int8_t board[8][8]);
+    bool memorize_board_state(const int board[8][8]);
     void clear_hashmap_history();
 private:
     struct BoardHasher {
-        std::size_t operator()(const int8_t board[8][8]) const;
+        std::size_t operator()(const int board[8][8]) const;
     };
 
     struct BoardEqual {
-        bool operator()(const int8_t lhs[8][8], const int8_t rhs[8][8]) const;
+        bool operator()(const int lhs[8][8], const int rhs[8][8]) const;
     };
 
     std::unordered_map<std::size_t, int> boardFrequency;
